Move Garak and Nessus configuration dataclasses to their new package location
